Rename blockchain and add contraint on blockchain roles
